extinderee logica storage

+++ src/services/carti-service.js
@@ -1,6 +1,6 @@
 export const CartiService = {
 	storageKey: "cartiData",
-	primaryKey: 0,
+	storageKeyPrimaryKey: "cartiPrimaryKey",
 
 	async getCartiInitial() {
 		let filePath = new URL("@/carti-db.json", import.meta.url).pathname;
@@ -12,8 +12,20 @@ export const CartiService = {
 			}
 
 			const data = await response.json();
-			localStorage.setItem(this.storageKey, JSON.stringify(data));
-			this.primaryKey = data.length > 0 ? Math.max(...data.map(carte => carte.id)) + 1 : 1;
+
+			// Extragem valorile din local storage
+			let existingData = localStorage.getItem(this.storageKey);
+			let existingPrimaryKey = localStorage.getItem(this.storageKeyPrimaryKey);
+
+			// Verificăm dacă avem valori în storage deja
+			if (!existingData) {
+				// Dacă nu avem, le adăugăm
+				localStorage.setItem(this.storageKey, JSON.stringify(data));
+			}
+
+			if (!existingPrimaryKey) {
+				localStorage.setItem(this.storageKeyPrimaryKey, JSON.stringify(data.length));
+			}
 		} catch (error) {
 			console.error("A apărut o eroare:", error);
 		}
@@ -25,17 +37,26 @@ export const CartiService = {
 		return carti.find(carte => carte.id === id);
 	},
 
+	cartiGetCount() {
+		return this.cartiSelect().length;
+	},
+
 	cartiSelect() {
 		let storedData = localStorage.getItem(this.storageKey);
 		return storedData ? JSON.parse(storedData) : [];
 	},
 
 	cartiInsert(carte) {
+		let primaryKey = this.primaryKeyGet();
+		primaryKey = primaryKey ? primaryKey + 1 : 1;
+
 		let storedData = localStorage.getItem(this.storageKey);
 		let carti = storedData ? JSON.parse(storedData) : [];
-		carte.id = this.primaryKey++;
+		carte.id = primaryKey;
 		carti.push(carte);
 		localStorage.setItem(this.storageKey, JSON.stringify(carti));
+
+		this.primaryKeyUpdate(primaryKey);
 	},
 
 	cartiUpdate(id, updatedCarte) {
@@ -53,5 +74,16 @@ export const CartiService = {
 		let carti = storedData ? JSON.parse(storedData) : [];
 		carti = carti.filter(carte => carte.id !== id);
 		localStorage.setItem(this.storageKey, JSON.stringify(carti));
+	},
+
+
+
+	primaryKeyGet() {
+		let primaryKey = localStorage.getItem(this.storageKeyPrimaryKey);
+		return primaryKey ? JSON.parse(primaryKey) : 0;
+	},
+
+	primaryKeyUpdate(newPrimaryKey) {
+		localStorage.setItem(this.storageKeyPrimaryKey, JSON.stringify(newPrimaryKey));
 	}
 };
